이미지 업로드 함수 통합: uploadProfileImage와 uploadProjectImage를 공통 uploadImage 함수로 리팩터링
